feat: File Icons in File Picker
Introduces optional file icons for the `FilePicker`. The icons are
accessed through the new `FileInfo` struct, which is a valid `Item`
that includes the icon in its `Span` (if configured to do so).

The idea is that `FileInfo` could be used for additional file type
metadata if it is needed in the future.

These icons assume the use of a nerd font, which is why they are
disabled by default.